moved config loading to separate file
