+++ pkg/config/config.go
@@ -1,13 +1,5 @@
 package config
 
-import (
-	"io"
-	"os"
-	"path/filepath"
-
-	"gopkg.in/yaml.v2"
-)
-
 // Config contains the configuration for the application
 type Config struct {
 	// DefaultFileType is the default file type (name) to use when creating a new entry
@@ -95,34 +87,4 @@ type Paths struct {
 type UserSettings struct {
 	// Timezone is the timezone to use for the application
 	Timezone string `yaml:"timezone,omitempty"`
-}
-
-// LoadConfig loads the configuration from a file
-func LoadConfig(configPath string) (*Config, error) {
-	config := &Config{}
-
-	file, err := os.Open(configPath)
-	if err != nil {
-		return nil, err
-	}
-	yamlData, err := io.ReadAll(file)
-	if err != nil {
-		return nil, err
-	}
-
-	if err := yaml.Unmarshal(yamlData, config); err != nil {
-		return nil, err
-	}
-
-	return config, nil
-}
-
-// GetDefaultConfigPath returns the default path to the configuration file
-func GetDefaultConfigPath() (string, error) {
-	home, err := os.UserHomeDir()
-	if err != nil {
-		return "", err
-	}
-	defaultConfigPath := filepath.Join(home, ".journal", "config.yaml")
-	return defaultConfigPath, nil
 }
